pyproject.toml
[project]
name = "inventory-comments"
version = "3.0.7"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic==1.8.0",
    "cryptography==39.0.0",
    "fastapi==0.95.2",
    "psycopg2-binary==2.9.6",
    "pyjwt==2.6.0",
    "requests==2.31.0",
    "sqlalchemy[asyncio]==1.4.54",
    "sqlmodel==0.0.11",
    "uvicorn==0.22.0",
]

[dependency-groups]
dev = [
]
lint = [
    "ruff==0.11.2",
]
tests = [
]
security = [
    "pip-audit==2.7.3",
]
migrations = [
    "alembic==1.8.0",
    "psycopg2-binary==2.9.6",
    "sqlalchemy==1.4.54",
    "sqlmodel==0.0.11",
]
